replace confirmation dialogs by toasts
The changes done when lading a demo or opening a file are undone with
the built-in undo mechanism in GtkSourceView.

Also the automatical running of javascript code after loading the demo
is now removed. It created issues because it prettified the code and
that would have been undone instead of the actual demo loading.