Simplify Library contrbutions
With this change, it's only needed to add the files to
`src/Library/demos`.

Fiddling with gresource, meson, ... is error prone and not very begineer
friendly.

We are now loading the demo files from disk instead of from a gresource.
It will be fine for now but we can optimize later when we have many
entries.