Separate JavaScript.
Some demo don't have code.
For Vala / Rust it works fine because we have a placeholder.

For JavaScript there is nothing in that case.
So it says reports as available with Rust/Vala but not JavaScript.

Most likely we would want to remove all tags for demos that don't have code but that's a bit more work.
Maybe later.