python: fix Previewer not being closable (#803)
This fixes #745.

Forgot a pretty critical "return" when I ported the Vala previewer,
which ran the code which was only supposed to run if a window was
already opened. This then immediately destroyed the just created and
shown window, leaving just a ghost window open :ghost:. Kinda impressed
Gtk still handled everything fine seemingly, except closing the window
again.
What is a bit strange though is that it doesn't even log a warning until
you try to close the window.

Also updated the docstring of the Previewer, the old docstring wasn't
accurate anymore.