Fix  error handling of LSP message reading.
LSPs, who exit early, trigger a lot of errors, because they do not send any headers.
This patch fixes the access of undefined headers to prevent runtime exceptions.

Furthermore, it adds a timeout to not overwhelm the garbage collection sweeping phase of GTK and adds debug output for non-JSON LSP responses.