Change the Python LSP to pylsp
As decided, this commit changes the executed LSP to pyslp and adds the missing code to it up.
Furthermore, it adds a missing return in the LSP's setup function, which doesn't seem to have triggered any issues, yet.

Although, a locally installed version of pylsp, initializes, it doesn't yet work, because it quit's after sending it's firsts message.