Fix blank preview on no code demos (#865)
Fixes https://github.com/workbenchdev/Workbench/issues/812
Closes https://github.com/workbenchdev/Workbench/pull/820

This is the culprit
https://github.com/workbenchdev/Workbench/blob/58b40d9802d14cbcc613a4f25489858783b808b2/src/window.js#L268-L272C6

For that particular code path (javascript selected, demo without js) it
wouldn't update the preview

One of the issue is that demos without code shouldn't have `autorun:
true`. I will fix that.
But in any case - I don't want Workbench to misbehave even if the demo
json is incorrect.