Compile TypeScript to JavaScript before executing (#941)
This basically compiles from TypeScript to JavaScript at runtime when
the "Run" button is clicked.

However, there are currently 2 issues worth mentioning:

### 1. Speed

Notice that this is noticeably slow because it's using `tsc`. It could
possibly be improved by using `esbuild`, `swc`, `babel` or something
similar but then there will be no typechecking when the "Run" button is
clicked.

However, I think the above typechecking caveat will not make much sense
when we have real-time Intellisense in the editor for TypeScript.

### 2. Sourcemaps

Another consideration is the lack of sourcemap support in GJS. While tsc
can generate sourcemaps, this feature is disabled because GJS won't use
them. This means that some errors will have the wrong line:column
information.

For example:


![image](https://github.com/workbenchdev/Workbench/assets/37999241/c6487292-18a9-4e50-85a0-5c8771f107fc)

Workbench/GJS reports the error is on line number 17 even if it's
actually on line number 22 because that's where it ends up after it's
compiled to JavaScript (many compilers will eat up unnecessary line
breaks even though the minify option is turned off).

I left some TODOs in here where some decisions need to be made and hope
to get some feedback